requirements.txt
Django==4.2.1
Pillow==9.5.0
django-admin-sortable2==2.1.8
django-tinymce==3.6.1